Create an SEIR model with contact matrix and ASF-inspired parameters

Creates a SimInf SEIRcm (siminf4movenet::SEIRcm()) model with the following fixed parameters relating to African swine fever:

epsilon_rate = 18/6.1

epsilon_shape = 18

gamma_rate = 1/8

gamma_shape = 1

N.B. This function requires the siminf4movenet package to be installed.

Usage

SEIRcm_ASF(infected, tspan, contact_matrix)

Arguments

infected

logical vector with the infectious status in each node.

tspan

A vector (length >= 1) of increasing time points where the state of each node is to be returned. Can be either an integer or a Date vector. A Date vector is coerced to a numeric vector as days, where tspan[1] becomes the day of the year of the first year of tspan. The dates are added as names to the numeric vector.

contact_matrix

an optional contact matrix to include between-node transmission of infection. The contact matrix specifies the strength of the interaction between nodes. Specifically, contact_matrix[i, j] measures the relative strength of transmission from node i to node j.

Details

Holding-level incubation rate (epsilon) related parameters

Parameters for Epsilon come from Guinat et al. 2018, via EFSA 2020. They are within-farm transmission parameters for one of 9 outbreaks of the Georgia 2007/1 strain in Russia. The rate and shape parameters are both close to the median for the 9 outbreaks. In the EFSA report, parameters from this outbreak are used as a low-R0 scenario. These parameters cover a larger range with a higher average, compared to several other studies - but this might fit with the idea that farm-level latent periods would be higher than individual-pig latent periods due to v low level transmissibility if considering just a single pig.

Holding-level removal rate (gamma) related parameters

Average time-to-removal (infectious period) was derived as follows: average time from first infection to suspicion report = 13 days (EFSA 2020) + assumed time from suspicion report to removal = 1 day - average latent period = 6 days (as from Guinat et al. 2018)

See also