Create a report that analyses how modifying movement weights affects network properties

create_anonymisation_effect_analysis_report() takes as input a livestock movement dataset, and analyses how network analyses of these data are affected by privacy-enhancing modifications (jittering or rounding) of movement weights. The function creates a report with visualisations of the effects of different amounts of jittering and rounding on a selection of epidemiologically relevant global network properties as well as on the ranking of holdings according to various centrality measures. The aim is to help livestock movement data managers and analysts find an appropriate balance between data privacy and their utility in network analyses.

Usage

create_anonymisation_effect_analysis_report(
  movement_data,
  output_file,
  modify_weights = TRUE,
  modify_dates = FALSE,
  n_jitter_sim = 3,
  time_unit = "28 days",
  verbose = FALSE
)

Arguments

movement_data

A movenet-format movement tibble.

output_file

Output file name and path.

modify_weights

A logical indicating whether to analyse the effects of modifying weights (default TRUE). Currently this MUST be TRUE for the function to proceed.

modify_dates

Does not currently do anything as analysis of modified dates is not yet implemented.

n_jitter_sim

An integer indicating the number of random jitter simulations.

time_unit

A character string indicating the time unit of analyses (default "28 days"). The overall movement data will be split into sub-networks based on this time unit.

verbose

A logical indicating whether to print progress messages (default FALSE).

Details

This function applies various privacy enhancing modifications to the provided movement_data, splits up these data into periods of the specified time_unit and converts them into networks, and then provides users with a report with network analyses comparing the true data with the modified data.

Prior to applying privacy enhancing modifications, movement data are pre-processed: repeated movements between the same holdings on the same day are aggregated, and self-moves and moves with weight 0 are removed.

Privacy enhancing modifications include the modification of movement weights (batch sizes) by jittering and rounding. For jittering, the function jitter_weights() is used, with jitter ranges of 5, 10, 50, ... up until the order of magnitude of the mean movement weight in the data. For rounding, the function round_weights() is used, with rounding units of 5, 10, 50, ... up until the order of magnitude of the largest movement weight in the data.

The function then creates static networks for the true data and for the data with modified weights.

It calculates a selection of global network properties for each network, and compares these properties between the true data and the modified data. It also calculates and compares the ranking of holdings according to various centrality measures between the true data and the modified data. The results are visualised in the report.