Create a networkDynamic edge.spells data frame from a movement data tibble

create_edge_spells() is an internal helper function that converts a modified movenet-format movement data tibble into a networkDynamic-compliant edge activity (edge.spells) data frame. It requires integer node identifiers, and dates converted to numeric format.

Usage

create_edge_spells(movement_data)

Arguments

movement_data

A movement data tibble with integer node identifiers and dates in numeric format.

Value

A networkDynamic-compliant edge.spells data frame, with 1 column more than movement_data. The first 4 columns of the returned data frame correspond to date, date again, from, and to; any additional movement_data columns are located thereafter.

See also