Create a vector of holdings infected at t=0 to plug into SimInf models, by randomly selecting a specified number of holdings

Create a vector of holdings infected at t=0 to plug into SimInf models, by randomly selecting a specified number of holdings

Usage

create_random_infected_vector(n_nodes, n_initially_infected)

Arguments

n_nodes

Number of nodes/holdings included in the model.

n_initially_infected

Number of holdings infected at t=0.

Value

A numeric vector, indicating which nodes are infected at t=0, to plug into the siminf4movenet::SEIRcm() model or another SimInf model.

See also