Create transmission probability matrix for SEIRcm model from movement and holding data

data2contactmatrix() creates a holding-to-holding transmission probability matrix for the siminf4movenet::SEIRcm() model. This matrix incorporates one or more of the following components:

Daily average movement-based transmission probabilities, calculated using the daily average weight moved between each pair of holdings (obtained from movement_data) and the provided weight_unit_transmission_probability.

Daily average local spread trasmission probabilities (optional), calculated using the distances between holdings (obtained from the coordinates in holding_data) and the provided local_spread_probabilities.

Any other user-provided transmission probability matrix (optional).

Usage

data2contactmatrix(
  movement_data,
  holding_data = NULL,
  incl_nonactive_holdings = TRUE,
  accept_missing_coordinates = FALSE,
  accept_missing_additional_tm_probabilities = FALSE,
  weight_unit_transmission_probability = 0.5,
  whole_months = TRUE,
  local_spread_transmission_probabilities = NULL,
  additional_transmission_prob_matrices = NULL
)

Arguments

movement_data

A movenet-format movement data tibble.

holding_data

A movenet-format holding data tibble (optional).

incl_nonactive_holdings

A logical indicating whether to include non-active holdings (default FALSE).

accept_missing_coordinates

A logical indicating whether to accept holding_data with missing coordinates (default FALSE). Any missing coordinates will result in an error if FALSE, or a warning and "local spread" (distance-based transmission) set to 0 in transmission matrix entries involving the holdings with missing coordinates if TRUE.

accept_missing_additional_tm_probabilities

A logical indicating whether to accept additional_transmission_prob_matrices with missing values (default FALSE). Any missing values will result in an error if FALSE, or a warning and the relevant transmission probability components set to 0 in transmission matrix entries involving the holdings with missing values if TRUE.

weight_unit_transmission_probability

A single numeric indicating transmission probability per unit weight moved (default 0.5)

whole_months

A logical indicating whether movement_data covers full months (default TRUE). This affects calculation of daily average weights, and thus of movement-based transmission probabilities. See @details.

local_spread_transmission_probabilities

A data frame or tibble of tiered local spread probabilities. This must consist of 3 columns: lower_boundary (in metres), upper_boundary (in metres), probability.

additional_transmission_prob_matrices

A named list with any additional transmission probability matrices to include.

Value

A named list with two elements:

transmission_matrix containing a matrix with overall daily transmission probabilities from each holding (each row) to each holding (each column), based on movements, (optionally) local spread, and (optionally) any other user-provided transmission probabilities between holdings. Holdings have been given new identifiers in the form of digits (format required by SimInf); the original identifiers can be looked up in the provided key. The matrix diagonal is set to 0, corresponding to no self-infection.

key containing a named character vector that links original holding identifiers (element names) and numeric node identifiers (as element values).

See also