Example movenet-format holding data

A fully synthetic livestock holding dataset, created to illustrate the structure of movenet-format holding data tibbles.

Usage

example_holding_data

Format

A data frame (tibble) with 500 rows and 4 columns:

cph

Unique holding identifier; a character string in the format of a UK County Parish Holding (CPH) number. This column corresponds to the id data field, required in movenet-format holding data.

holding_type

Type of holding (character string). This column corresponds to the type data field, optional in movenet-format holding data.

herd_size

Herd size, in numeric (double) format. This column corresponds to the herd_size data field, optional in movenet-format holding data.

coordinates

Holding coordinates in degrees longitude and latitude, using the ETRS89 coordinate reference system, and formated as a simple feature (sf) geometry list-column (class sfc_POINT). This column is derived from coord_x and coord_y data fields, optional in movenet-format holding data.

Source

The content of this dataset was generated by random number and text generation, using UK-inspired data formats (e.g. County Parish Holding (CPH) identifiers). It does not correspond to or derive from any real holding data.

The csv-format data file that this holding data tibble was derived from is accessible at system.file("extdata/example_holding_data.csv", package = "movenet"). The tibble was obtained by calling reformat_data() on this csv file.

See also