Replace holding identifiers with consecutive "integers" (in character format) starting from 1, and harmonise movement and holding data

holdingids2consecints() is an internal helper function that replaces holding identifiers in movement_data and (optional) holding_data with consecutive "integers" (in character format), ranging from 1 to the total number of node identifiers.

Additionally, if holding_data is provided, it checks whether the included holding identifiers correspond to those in movement_data, and it harmonises both data tibbles:

If movement_data contains additional holdings, entries for these are added to holding_data, with any columns beyond id being assigned the value NA.

If holding_data contains any entries for holdings that are not in movement_data (i.e. entries for non-active holdings), these are removed if incl_nonactive_holdings == FALSE or kept otherwise.

This function facilitates conversion of movenet data tibbles into formats compatible with networkDynamic::networkDynamic() and SimInf::SimInf, which require node identifiers to be consecutive integers starting from 1.

Usage

holdingids2consecints(
  movement_data,
  holding_data = NULL,
  incl_nonactive_holdings = FALSE
)

Arguments

movement_data

A movenet-format movement data tibble.

holding_data

A movenet-format holding data tibble (optional).

incl_nonactive_holdings

A logical that indicates whether to keep entries in holding_data for holdings that are not present in movement_data. Default is FALSE.

Value

A named list with 3 elements:

key: A named character vector representing the identifier replacement key. Names are original identifiers, values are new integer identifiers.

movement_data: movement_data with holding identifiers replaced with integers.

holding_data: Either holding_data with holding identifiers replaced with integers, and potentially some additional or removed rows; or NULL if no holding_data was provided.